refactor: icon public 으로 이동
