--- components/Header/icon.tsx
@@ -1,34 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const LoginIcon = createIcon({
-  displayName: 'LoginIcon',
-  viewBox: '0 0 40 40',
-  path: (
-    <path
-      d="M18.0312 11.6894L15.7808 13.8642L19.9601 17.903H3.56445V21.0098H19.9601L15.7808 25.0486L18.0312 27.2234L26.0683 19.4564L18.0312 11.6894ZM32.498 30.3302H19.6386V33.437H32.498C34.2662 33.437 35.7128 32.0389 35.7128 30.3302V8.58263C35.7128 6.87389 34.2662 5.47583 32.498 5.47583H19.6386V8.58263H32.498V30.3302Z"
-      fill="#F96400"
-    />
-  ),
-});
-
-export const LogoutIcon = createIcon({
-  displayName: 'LogoutIcon',
-  viewBox: '0 0 63 63',
-  path: (
-    <path
-      d="M45.3333 18.6667L41.5733 22.4267L48.4533 29.3333H21.3333V34.6667H48.4533L41.5733 41.5467L45.3333 45.3333L58.6666 32L45.3333 18.6667ZM10.6666 13.3333H31.9999V8H10.6666C7.73325 8 5.33325 10.4 5.33325 13.3333V50.6667C5.33325 53.6 7.73325 56 10.6666 56H31.9999V50.6667H10.6666V13.3333Z"
-      fill="#F96400"
-    />
-  ),
-});
-
-export const UserIcon = createIcon({
-  displayName: 'UserIcon',
-  viewBox: '0 0 63 63',
-  path: (
-    <path
-      d="M31.5 31.5C37.3013 31.5 42 26.8013 42 21C42 15.1988 37.3013 10.5 31.5 10.5C25.6987 10.5 21 15.1988 21 21C21 26.8013 25.6987 31.5 31.5 31.5ZM31.5 36.75C24.4913 36.75 10.5 40.2675 10.5 47.25V52.5H52.5V47.25C52.5 40.2675 38.5088 36.75 31.5 36.75Z"
-      fill="#F96400"
-    />
-  ),
-});
